test: add success and failure tests for all viewmodel methods and handle network/empty view errors
- Added unit tests to verify success and failure cases for all methods in viewmodels
- Implemented comprehensive error handling for network-related issues
- Added logic to handle empty states in views to improve user experience
- Utilized mock data and dependencies to simulate API responses
- Ensured proper error messages and states are displayed for network errors and empty views
- Enhanced test coverage to ensure robustness and reliability of viewmodels and UI components

+++ SnapShop-Admin/Views/Collections/CollectionsView.swift
@@ -23,38 +23,42 @@ struct CollectionsView: View {
                             collectionsViewModel.getCollections()
                         }
                 }else{
-                    ScrollView{
-                        CollectionsGrid(screenHeight: geometry.size.height, collections: collectionsViewModel.collections, deleteCollection: { collection in
-                            selectedCollection = collection
-                        }, viewModel: collectionsViewModel)
-                    }
-                    .navigationBarTitle("Collections")
-                    .toolbar {
-                        ToolbarItem(placement: .navigationBarTrailing) {
-                            Button{
-                                openAddCollectionView.toggle()
-                            }label: {
-                                Image(systemName: "plus.app")
-                                    .font(.system(size: 24))
-                            }
-                            .sheet(isPresented: $openAddCollectionView){
-                                AddCollectionView(collectionsViewModel: collectionsViewModel)
-                            }
-                            
+                    if collectionsViewModel.collections.isEmpty {
+                        ContentUnavailableView(title: "No collections added yet!", imageName: "rectangle.stack.badge.plus")
+                    }else{
+                        ScrollView{
+                            CollectionsGrid(screenHeight: geometry.size.height, collections: collectionsViewModel.collections, deleteCollection: { collection in
+                                selectedCollection = collection
+                            }, viewModel: collectionsViewModel)
                         }
-                    }.alert(item: $selectedCollection) { collection in
-                        Alert(
-                            title: Text("Confirm Deletion"),
-                            message: Text("Are you sure you want to delete this collection?"),
-                            primaryButton: .destructive(Text("Delete")) {
-                                collectionsViewModel.deleteCollection(collection: collection)
-                            },
-                            secondaryButton: .cancel(Text("Cancel"))
-                        )
+                        .navigationBarTitle("Collections")
+                        .toolbar {
+                            ToolbarItem(placement: .navigationBarTrailing) {
+                                Button{
+                                    openAddCollectionView.toggle()
+                                }label: {
+                                    Image(systemName: "plus.app")
+                                        .font(.system(size: 24))
+                                }
+                                .sheet(isPresented: $openAddCollectionView){
+                                    AddCollectionView(collectionsViewModel: collectionsViewModel)
+                                }
+                                
+                            }
+                        }.alert(item: $selectedCollection) { collection in
+                            Alert(
+                                title: Text("Confirm Deletion"),
+                                message: Text("Are you sure you want to delete this collection?"),
+                                primaryButton: .destructive(Text("Delete")) {
+                                    collectionsViewModel.deleteCollection(collection: collection)
+                                },
+                                secondaryButton: .cancel(Text("Cancel"))
+                            )
+                        }
+                        
                     }
-                   
                 }
-            }
+            }.showAlert(for: $collectionsViewModel.userError)
         }
     }
 }
